pyproject.toml
[project]
name = "truefact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "masumi>=0.1.34",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
